requirements.txt
accelerate==0.28.0
Cython==3.0.10
datasets==2.17.0
faiss-gpu==1.7.2
json-lines==0.5.0
nltk==3.8.1
numpy==1.24.3
openai==0.27.8
packaging==23.1
pandas==2.0.3
peft==0.10.0
Pillow==9.5.0
protobuf==3.20.1
psutil==5.9.5
pyjnius==1.6.1
pyserini==0.22.1
PyYAML==6.0
rank-bm25==0.2.2
regex==2023.5.5
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
spacy==3.5.3
tokenizers==0.15.2
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.65.0
transformers==4.38.1
triton==2.2.0
trl==0.8.1
wandb==0.16.3
wheel==0.41.2